config.py
import os
from dotenv import load_dotenv

base = os.path.abspath(os.path.dirname(__name__))
load_dotenv(os.path.join(base, '.env'))

class Config:
    FLASK_APP = os.getenv('FLASK_APP')
    FLASK_ENV = os.getenv('FLASK_ENV')
    SECRET_KEY = os.getenv('SECRET_KEY')
    SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_DATABASE_URI')